requirements.txt
certbot>=2.7.0
acme>=2.7.0
cryptography>=41.0.0
requests>=2.31.0
python-dateutil>=2.8.2
josepy>=1.14.0
pytz>=2023.3
pyyaml>=6.0
